pyproject.toml
[project]
name = "kajisuke-django"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "jpholiday (>=1.0.2,<2.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^5.2.0"

